+++ src/views/translate-message.jsx
@@ -1,28 +1,32 @@
-var _ = require('lodash');
 var React = require('react');
 var CreateReactClass = require('create-react-class');
 var ReactDOMServer = require('react-dom-server');
 var Translate = require('../translate.js');
 
 var TranslateMessage = CreateReactClass({
-    render: function () {
-        var values = this.props.values;
-        _.each(values, function (value, index) {
-            if (React.isValidElement(value)) {
-                values[index] = ReactDOMServer.renderToStaticMarkup(value);
-            }
-        });
+  render: function () {
+      const { values, message, doubleSpan } = this.props;
 
-        if (this.props.doubleSpan) {
-            return <span>
-                       <span className={ 'translate-message' }
-                          dangerouslySetInnerHTML={{__html: Translate.translate(this.props.message, {amount: ''})}}/>
-                       <span className={ 'translate-message-amount' } dangerouslySetInnerHTML={{__html: values.amount}}/>
-                   </span>
-        } else {
-            return <span className={ 'translate-message' }
-                         dangerouslySetInnerHTML={{__html: Translate.translate(this.props.message, values)}}/>;
-        }
-    }
+      const translateContent = (values && Object.keys(values)
+          .filter(key => React.isValidElement(values[key]))
+          .reduce((acc, curr) => {
+              acc[curr] = ReactDOMServer.renderToStaticMarkup(values[curr]);
+              return acc;
+          }, {})) || [];
+
+      if (doubleSpan) {
+          return <span>
+                      <span className={ 'translate-message' }
+                        dangerouslySetInnerHTML={{__html: Translate.translate(message, {amount: ''})}}/>
+                      <span
+                          className={ 'translate-message-amount' }
+                          dangerouslySetInnerHTML={{__html: translateContent.amount}}
+                      />
+                  </span>
+      } else {
+          return <span className={ 'translate-message' }
+                        dangerouslySetInnerHTML={{__html: Translate.translate(message, translateContent)}}/>;
+      }
+  }
 });
 module.exports = TranslateMessage;